+++ src/main/java/com/transcotech/transcota_system/Service/DriverServiceInterface.java
@@ -7,8 +7,8 @@ import com.transcotech.transcota_system.model.User;
 public interface DriverServiceInterface {
 
     public List<User> findAll();
-    public User searchId(Integer id);
-    public void deleteUser(Integer id);
+    public User searchId(Long id);
+    public void deleteUser(Long id);
     public User createDriver(User user);
 
 }
